third_party/blink/web_tests/fast/dom/NamedNodeMap-missing-arguments-expected.txt

PASS attributes.getNamedItem() threw exception TypeError: Failed to execute 'getNamedItem' on 'NamedNodeMap': 1 argument required, but only 0 present..
PASS attributes.getNamedItemNS() threw exception TypeError: Failed to execute 'getNamedItemNS' on 'NamedNodeMap': 2 arguments required, but only 0 present..
PASS attributes.getNamedItemNS('http://www.w3.org/2000/svg') threw exception TypeError: Failed to execute 'getNamedItemNS' on 'NamedNodeMap': 2 arguments required, but only 1 present..
PASS attributes.item() threw exception TypeError: Failed to execute 'item' on 'NamedNodeMap': 1 argument required, but only 0 present..
PASS attributes.removeNamedItem() threw exception TypeError: Failed to execute 'removeNamedItem' on 'NamedNodeMap': 1 argument required, but only 0 present..
PASS attributes.removeNamedItemNS() threw exception TypeError: Failed to execute 'removeNamedItemNS' on 'NamedNodeMap': 2 arguments required, but only 0 present..
PASS attributes.removeNamedItemNS('http://www.w3.org/2000/svg') threw exception TypeError: Failed to execute 'removeNamedItemNS' on 'NamedNodeMap': 2 arguments required, but only 1 present..
PASS attributes.setNamedItem() threw exception TypeError: Failed to execute 'setNamedItem' on 'NamedNodeMap': 1 argument required, but only 0 present..
PASS attributes.setNamedItemNS() threw exception TypeError: Failed to execute 'setNamedItemNS' on 'NamedNodeMap': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
